added outline for login
